[sign-in gate] split test file
