+++ src/server/signin-gate/libPureTests/articleIdentifierIsAllowed.test.ts
@@ -1,0 +1,12 @@
+import { articleIdentifierIsAllowed } from '../libPure';
+
+it('articleIdentifierIsAllowed', () => {
+    expect(
+        articleIdentifierIsAllowed(
+            'www.theguardian.com/money/2017/mar/10/ministers-to-criminalise-use-of-ticket-tout-harvesting-software',
+        ),
+    ).toBe(true);
+    expect(articleIdentifierIsAllowed('www.theguardian.com/tips')).toBe(false);
+    expect(articleIdentifierIsAllowed('www.theguardian.com/tips#test')).toBe(false);
+    expect(articleIdentifierIsAllowed('www.theguardian.com/tips/test')).toBe(false);
+});
