+++ src/server/signin-gate/libPureTests/getTreatmentsRequestPayloadToGateType.test.ts
@@ -1,325 +1,7 @@
-import {
-    articleIdentifierIsAllowed,
-    buildGetTreatmentsRequestPayload,
-    buildGuUserTreatmentsEnvelop,
-    buildLogTreatmentInteractionRequestPayload,
-    getTreatmentsRequestPayloadToGateType,
-    guDismissibleUserTreatment,
-    guMandatoryUserTreatment,
-    isValidContentType,
-    isValidSection,
-    isValidTagIdCollection,
-    mvtIdIsAuxiaAudienceShare,
-    userTreatmentsEnvelopToProxyGetTreatmentsAnswerData,
-} from '../libPure';
+import { getTreatmentsRequestPayloadToGateType } from '../libPure';
 import type { GetTreatmentsRequestPayload } from '../types';
 
-describe('buildGetTreatmentsRequestPayload', () => {
-    it('should return return the right payload', () => {
-        const projectId = 'projectId';
-        const browserId = 'browserId';
-        const isSupporter = true;
-        const dailyArticleCount = 21;
-        const articleIdentifier = 'articleIdentifier';
-        const editionId = 'UK';
-        const countryCode = 'GB';
-        const hasConsented = true;
-        const shouldServeDismissible = false;
-
-        const expectedAnswer = {
-            projectId,
-            userId: browserId,
-            contextualAttributes: [
-                {
-                    key: 'is_supporter',
-                    boolValue: isSupporter,
-                },
-                {
-                    key: 'daily_article_count',
-                    integerValue: dailyArticleCount,
-                },
-                {
-                    key: 'article_identifier',
-                    stringValue: articleIdentifier,
-                },
-                {
-                    key: 'edition',
-                    stringValue: editionId,
-                },
-                {
-                    key: 'country_key',
-                    stringValue: countryCode,
-                },
-                {
-                    key: 'has_consented',
-                    boolValue: hasConsented,
-                },
-                {
-                    key: 'should_not_serve_mandatory',
-                    boolValue: shouldServeDismissible,
-                },
-            ],
-            surfaces: [
-                {
-                    surface: 'ARTICLE_PAGE',
-                    maximumTreatmentCount: 1,
-                },
-            ],
-            languageCode: 'en-GB',
-        };
-
-        const returnedAnswer = buildGetTreatmentsRequestPayload(
-            projectId,
-            browserId,
-            isSupporter,
-            dailyArticleCount,
-            articleIdentifier,
-            editionId,
-            countryCode,
-            hasConsented,
-            shouldServeDismissible,
-        );
-        expect(returnedAnswer).toStrictEqual(expectedAnswer);
-    });
-});
-
-describe('guDefaultGateGetTreatmentsResponseData', () => {
-    const dismissibleGuGateTreatment = guDismissibleUserTreatment();
-    const mandatoryGuGateTreatment = guMandatoryUserTreatment();
-
-    it('should not return gate data if the number of gate dismissal is more than 5 (low gate display count)', () => {
-        const gateDismissCount = 6;
-        const gateDisplayCount = 0; // dismissible gate
-        const expectAnswer = {
-            responseId: '',
-            userTreatments: [],
-        };
-        expect(
-            buildGuUserTreatmentsEnvelop(gateDismissCount, gateDisplayCount, 'IE'),
-        ).toStrictEqual(expectAnswer);
-        expect(
-            buildGuUserTreatmentsEnvelop(gateDismissCount, gateDisplayCount, 'FR'),
-        ).toStrictEqual(expectAnswer);
-    });
-
-    it('should not return gate data if the number of gate dismissal is more than 5 (high gate display count)', () => {
-        const gateDismissCount = 6;
-        const gateDisplayCount = 10; // mandatory gate
-        const expectAnswer = {
-            responseId: '',
-            userTreatments: [],
-        };
-        expect(
-            buildGuUserTreatmentsEnvelop(gateDismissCount, gateDisplayCount, 'IE'),
-        ).toStrictEqual(expectAnswer);
-        expect(
-            buildGuUserTreatmentsEnvelop(gateDismissCount, gateDisplayCount, 'FR'),
-        ).toStrictEqual(expectAnswer);
-    });
-
-    it('[in Ireland] should return a dismissible gate if gateDisplayCount is in {0, 1, 2}', () => {
-        const gateDismissCount = 2; // low number allowing for a gate
-        const gateDisplayCount = 0; // dismissible gate
-        const expectAnswer = {
-            responseId: '',
-            userTreatments: [dismissibleGuGateTreatment],
-        };
-        expect(
-            buildGuUserTreatmentsEnvelop(gateDismissCount, gateDisplayCount, 'IE'),
-        ).toStrictEqual(expectAnswer);
-    });
-
-    it('[in Ireland] should return a dismissible gate if gateDisplayCount is in {0, 1, 2}', () => {
-        const gateDismissCount = 2; // low number allowing for a gate
-        const gateDisplayCount = 2; // dismissible gate
-        const expectAnswer = {
-            responseId: '',
-            userTreatments: [dismissibleGuGateTreatment],
-        };
-        expect(
-            buildGuUserTreatmentsEnvelop(gateDismissCount, gateDisplayCount, 'IE'),
-        ).toStrictEqual(expectAnswer);
-    });
-
-    it('[in Ireland] should return a mandatory gate if gateDisplayCount is >= 3', () => {
-        const gateDismissCount = 2; // low number allowing for a gate
-        const gateDisplayCount = 3; // mandatory gate
-        const expectAnswer = {
-            responseId: '',
-            userTreatments: [mandatoryGuGateTreatment],
-        };
-        expect(
-            buildGuUserTreatmentsEnvelop(gateDismissCount, gateDisplayCount, 'IE'),
-        ).toStrictEqual(expectAnswer);
-    });
-
-    it('[Ireland] should return a mandatory gate if gateDisplayCount is >= 3', () => {
-        const gateDismissCount = 2; // low number allowing for a gate
-        const gateDisplayCount = 6; // mandatory gate
-        const expectAnswer = {
-            responseId: '',
-            userTreatments: [mandatoryGuGateTreatment],
-        };
-        expect(
-            buildGuUserTreatmentsEnvelop(gateDismissCount, gateDisplayCount, 'IE'),
-        ).toStrictEqual(expectAnswer);
-    });
-
-    it('[outside Ireland] should return a dismissible gate for any gateDisplayCount', () => {
-        const gateDismissCount = 2; // low number allowing for a gate
-        const gateDisplayCount = 0; // mandatory gate
-        const expectAnswer = {
-            responseId: '',
-            userTreatments: [dismissibleGuGateTreatment],
-        };
-        expect(
-            buildGuUserTreatmentsEnvelop(gateDismissCount, gateDisplayCount, 'FR'),
-        ).toStrictEqual(expectAnswer);
-    });
-
-    it('[outside Ireland] should return a dismissible gate for any gateDisplayCount', () => {
-        const gateDismissCount = 2; // low number allowing for a gate
-        const gateDisplayCount = 6; // mandatory gate
-        const expectAnswer = {
-            responseId: '',
-            userTreatments: [dismissibleGuGateTreatment],
-        };
-        expect(
-            buildGuUserTreatmentsEnvelop(gateDismissCount, gateDisplayCount, 'FR'),
-        ).toStrictEqual(expectAnswer);
-    });
-});
-
-describe('isValidContentType', () => {
-    it('accepts Article', () => {
-        expect(isValidContentType('Article')).toBe(true);
-    });
-
-    it('does not accepts NonArticle', () => {
-        expect(isValidContentType('NonArticle')).toBe(false);
-    });
-});
-
-describe('isValidSection', () => {
-    it('accepts news', () => {
-        expect(isValidSection('news')).toBe(true);
-    });
-
-    it('does not accept `about`', () => {
-        // `about` is taken from the list of hard coded invalid sections
-        expect(isValidSection('about')).toBe(false);
-    });
-});
-
-describe('isValidTagIdCollection', () => {
-    it('accepts `random`', () => {
-        expect(isValidTagIdCollection(['random/random', 'random/otherRandom'])).toBe(true);
-    });
-
-    it('does not accept `info/newsletter-sign-up`', () => {
-        // `info/newsletter-sign-up` is taken from the list of hard coded invalid sections
-        expect(isValidTagIdCollection(['info/newsletter-sign-up', 'random/otherRandom'])).toBe(
-            false,
-        );
-    });
-});
-
-describe('buildAuxiaProxyGetTreatmentsResponseData', () => {
-    it('build things correctly, in the case of a provided treatment', () => {
-        const auxiaData = {
-            responseId: 'responseId',
-            userTreatments: [
-                {
-                    treatmentId: 'treatmentId',
-                    treatmentTrackingId: 'treatmentTrackingId',
-                    rank: 'rank',
-                    contentLanguageCode: 'contentLanguageCode',
-                    treatmentContent: 'treatmentContent',
-                    treatmentType: 'treatmentType',
-                    surface: 'surface',
-                },
-            ],
-        };
-        const expectedAnswer = {
-            responseId: 'responseId',
-            userTreatment: {
-                treatmentId: 'treatmentId',
-                treatmentTrackingId: 'treatmentTrackingId',
-                rank: 'rank',
-                contentLanguageCode: 'contentLanguageCode',
-                treatmentContent: 'treatmentContent',
-                treatmentType: 'treatmentType',
-                surface: 'surface',
-            },
-        };
-        expect(userTreatmentsEnvelopToProxyGetTreatmentsAnswerData(auxiaData)).toStrictEqual(
-            expectedAnswer,
-        );
-    });
-
-    it('build things correctly, in the case of no treatment', () => {
-        const auxiaData = {
-            responseId: 'responseId',
-            userTreatments: [],
-        };
-        const expectedAnswer = {
-            responseId: 'responseId',
-            userTreatment: undefined,
-        };
-        expect(userTreatmentsEnvelopToProxyGetTreatmentsAnswerData(auxiaData)).toStrictEqual(
-            expectedAnswer,
-        );
-    });
-});
-
-describe('buildLogTreatmentInteractionRequestPayload', () => {
-    it('', () => {
-        const expectedAnswer = {
-            projectId: 'projectId',
-            userId: 'browserId',
-            treatmentTrackingId: 'treatmentTrackingId',
-            treatmentId: 'treatmentId',
-            surface: 'surface',
-            interactionType: 'interactionType',
-            interactionTimeMicros: 123456789,
-            actionName: 'actionName',
-        };
-        expect(
-            buildLogTreatmentInteractionRequestPayload(
-                'projectId',
-                'browserId',
-                'treatmentTrackingId',
-                'treatmentId',
-                'surface',
-                'interactionType',
-                123456789,
-                'actionName',
-            ),
-        ).toStrictEqual(expectedAnswer);
-    });
-});
-
-describe('articleIdentifierIsAllowed', () => {
-    expect(
-        articleIdentifierIsAllowed(
-            'www.theguardian.com/money/2017/mar/10/ministers-to-criminalise-use-of-ticket-tout-harvesting-software',
-        ),
-    ).toBe(true);
-    expect(articleIdentifierIsAllowed('www.theguardian.com/tips')).toBe(false);
-    expect(articleIdentifierIsAllowed('www.theguardian.com/tips#test')).toBe(false);
-    expect(articleIdentifierIsAllowed('www.theguardian.com/tips/test')).toBe(false);
-});
-
-describe('mvtIdIsInTheNaturalAuxiaShareOfAudience', () => {
-    expect(mvtIdIsAuxiaAudienceShare(0)).toBe(false);
-    expect(mvtIdIsAuxiaAudienceShare(1)).toBe(true);
-    expect(mvtIdIsAuxiaAudienceShare(210945)).toBe(true);
-    expect(mvtIdIsAuxiaAudienceShare(210946)).toBe(true);
-    expect(mvtIdIsAuxiaAudienceShare(350000)).toBe(true);
-    expect(mvtIdIsAuxiaAudienceShare(350001)).toBe(false);
-});
-
-describe('decideGateTypeFromGetTreatmentsRequestPayload', () => {
+describe('getTreatmentsRequestPayloadToGateType', () => {
     it('test shouldServeDismissible and showDefaultGate:mandatory interacting together', () => {
         const body: GetTreatmentsRequestPayload = {
             browserId: 'sample',
